requirements.txt
Bio
requests
pandas
pyyaml